Corrigindo tratamento de exceções

+++ workspace/NgmAdota/src/main/java/com/example/NgmAdota/modules/usuario/UsuarioRepository.java
@@ -2,6 +2,7 @@ package com.example.NgmAdota.modules.usuario;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
     UsuarioModel findByEmail(String email);
+
 }
